Clean up package structure
